Move traits to separate folder
