+++ include/CPlayTime.h
@@ -1,23 +1,12 @@
 #pragma once
 
 #include "IJsonable.h"
+#include "traits/CPlayTime.h"
 
 #include "types/CPlayTime.h"
 
 namespace futsim
 {
-
-class CPlayTime;
-
-template <> struct json_traits<CPlayTime>
-{
-	//! JSON key for the class.
-	static inline constexpr std::string_view KEY = "Play time";
-	//! JSON key for the \copybrief CPlayTime::mPeriodCount
-	static inline constexpr std::string_view PERIOD_COUNT_KEY = "Period count";
-	//! JSON key for the \copybrief CPlayTime::mPeriodTime
-	static inline constexpr std::string_view PERIOD_TIME_KEY = "Period time";
-};
 
 /**
  * @brief Class that defines a play time phase of a match.
