+++ include/football/CMatch.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "IJsonable.h"
+#include "traits/json.h"
 
 #include "football/CTeam.h"
 #include "football/CStadium.h"
