+++ include/football/CMatchState.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "IJsonable.h"
+#include "traits/json.h"
 
 #include "types/CMatchState.h"
