+++ include/football/CPossessionDrawConfiguration.h
@@ -1,10 +1,11 @@
 #pragma once
 
 #include "IJsonable.h"
+#include "traits/default.h"
+#include "traits/json.h"
 
 #include "football/types/CDrawConfiguration.h"
 #include "football/types/CPossessionDrawConfiguration.h"
-#include "DefaultTraits.h"
 
 namespace futsim
 {
