+++ include/football/CTeam.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "IJsonable.h"
+#include "traits/json.h"
 
 #include "football/CPlayer.h"
 #include "football/types/CTeam.h"
