+++ include/traits/CPlayTime.h
@@ -1,0 +1,22 @@
+#pragma once
+
+#include "IJsonable.h"
+
+#include "traits/json.h"
+
+namespace futsim
+{
+
+class CPlayTime;
+
+template <> struct json_traits<CPlayTime>
+{
+	//! JSON key for the class.
+	static inline constexpr std::string_view KEY = "Play time";
+	//! JSON key for the \copybrief CPlayTime::mPeriodCount
+	static inline constexpr std::string_view PERIOD_COUNT_KEY = "Period count";
+	//! JSON key for the \copybrief CPlayTime::mPeriodTime
+	static inline constexpr std::string_view PERIOD_TIME_KEY = "Period time";
+};
+
+} // futsim namespace
